requirements.txt
# local package
-e .

# # external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1


absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
blitz-bayesian-pytorch==0.2.8
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dm-tree==0.1.8
edward2==0.0.2
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2024.2.0
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
hydra-core==1.3.2
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdock==0.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.9
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtAds==3.8.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semver==3.0.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.23.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wandb==0.16.2
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1